+++ packages/server/src/auth/CrossmintAuth.ts
@@ -1,0 +1,34 @@
+import { type Crossmint, CrossmintApiClient } from "@crossmint/common-sdk-base";
+import { CROSSMINT_API_VERSION, SDK_NAME, SDK_VERSION } from "./utils/constants";
+
+export class CrossmintAuth {
+    private crossmint: Crossmint;
+    private apiClient: CrossmintApiClient;
+
+    private constructor(crossmint: Crossmint) {
+        this.crossmint = crossmint;
+        this.apiClient = new CrossmintApiClient(this.crossmint, {
+            internalConfig: {
+                sdkMetadata: {
+                    name: SDK_NAME,
+                    version: SDK_VERSION,
+                },
+            },
+        });
+    }
+
+    public static from(crossmint: Crossmint): CrossmintAuth {
+        return new CrossmintAuth(crossmint);
+    }
+
+    async getUser(externalUserId: string) {
+        const result = await this.apiClient.get(`api/${CROSSMINT_API_VERSION}/sdk/auth/user/${externalUserId}`, {
+            headers: {
+                "Content-Type": "application/json",
+            },
+        });
+
+        const user = await result.json();
+        return user;
+    }
+}