+++ packages/server/src/auth/utils/constants.ts
@@ -1,0 +1,5 @@
+import { version, name } from "../../../package.json";
+
+export const SDK_VERSION = version;
+export const SDK_NAME = name;
+export const CROSSMINT_API_VERSION = "2024-09-26";